+++ pages/chat/[cid].tsx
@@ -1,6 +1,5 @@
 import styled from '@emotion/styled';
 import { useRouter } from 'next/router';
-import { ChatCompletionRequestMessage, Configuration, OpenAIApi } from 'openai';
 import { KeyboardEvent, useEffect, useRef, useState } from 'react';
 import { BsSend } from 'react-icons/bs';
 import { useRecoilValue } from 'recoil';
@@ -11,15 +10,8 @@ import Button from '@/components/common/Button';
 import InputText from '@/components/common/InputText';
 import Header from '@/components/Header';
 import { COLOR } from '@/constants/color';
-import { Message } from '@/types/message';
+import useChat from '@/hooks/useChat';
 import { RoomInfo } from '@/types/roomInfo';
-
-const MAKE_HEADER = (maxNumber: number): ChatCompletionRequestMessage => {
-  return {
-    role: 'system',
-    content: `You are chatbots in a chat room where one user and ${maxNumber} chatbots converse. Each message is in the format of number: content, where 0 represents the user's message and 1 to n represents the messages of the nth chatbot. You are responsible for assuming the role of all 1 to ${maxNumber} chatbots and participating actively in the conversation. Each chatbot can speak all at once, only a few at a time, or one chatbot can speak multiple times. Please format your response in number: content format, with each message separated by two line breaks.`,
-  };
-};
 
 const Chat = () => {
   const router = useRouter();
@@ -27,14 +19,16 @@ const Chat = () => {
   const API_KEY = useRecoilValue(atomApiKey);
 
   const [locationState, setLocationState] = useState<number>(0); // 0 시작, 1 유효, 2 오류
-  const [messages, setMessages] = useState<Message[]>([]);
   const [input, setInput] = useState<string>('');
-  const [isPending, setIsPending] = useState<boolean>(false);
+
   const [roomInfo, setRoomInfo] = useState<RoomInfo>({
     id: -1,
     name: '',
     people: 0,
   });
+  const { isPending, messages, setMessages, submitChat } = useChat(
+    roomInfo.people,
+  );
   const chatListRef = useRef<HTMLDivElement>(null);
 
   useEffect(() => {
@@ -65,90 +59,10 @@ const Chat = () => {
     setLocationState(1);
   }, [cid]);
 
-  const getNewId = () => {
-    if (!messages.length) return 1;
-    return messages[messages.length - 1].id + 1;
-  };
-
-  const doChat = async (message: string) => {
-    const config = new Configuration({
-      apiKey: API_KEY,
-    });
-    delete config.baseOptions.headers['User-Agent'];
-    const openai = new OpenAIApi(config);
-    const completion = await openai.createChatCompletion({
-      model: 'gpt-3.5-turbo',
-      messages: [
-        MAKE_HEADER(roomInfo.people - 1),
-        ...chatHistory(),
-        { role: 'user', content: message },
-      ],
-    });
-
-    return completion;
-  };
-
-  const chatHistory = () => {
-    const COUNT = roomInfo.people * 2;
-    const rawMessages = messages.slice(-COUNT);
-    const msgs: ChatCompletionRequestMessage[] = rawMessages.map((m) => {
-      return {
-        role: m.author === 0 ? 'user' : 'assistant',
-        content: `${m.author}: ${m.content}`,
-      };
-    });
-    return msgs;
-  };
-
   const onKeydown = (e: KeyboardEvent<HTMLInputElement>) => {
     const { key } = e;
     if (key !== 'Enter') return;
-    submitChat();
-  };
-
-  const submitChat = () => {
-    if (isPending || !input.length) return;
-    setIsPending(true);
-    const msg: Message = {
-      author: 0,
-      content: input,
-      createdAt: Date.now(),
-      id: getNewId(),
-    };
-    doChat(input).then((r) => {
-      console.log(r.data);
-      if (
-        r.data.choices[0].finish_reason !== 'stop' ||
-        !r.data.choices[0].message
-      ) {
-        const msg: Message = {
-          author: 1,
-          content: '오류가 발생했습니다. 다시 한 번 시도해 주세요.',
-          createdAt: Date.now(),
-          id: getNewId(),
-        };
-        setMessages((prev) => [...prev, msg]);
-        setIsPending(false);
-        return;
-      }
-      const msgs = r.data.choices[0].message.content
-        .split(/\n+(?=[1-4])/)
-        .map((c) => {
-          const sp = c.split(': ');
-          const author = parseInt(sp[0]);
-          const content = sp.slice(1).join(': ');
-          const msg: Message = {
-            author: author,
-            content: content,
-            createdAt: Date.now(),
-            id: getNewId() + author,
-          };
-          return msg;
-        });
-      setMessages((prev) => [...prev, ...msgs]);
-      setIsPending(false);
-    });
-    setMessages([...messages, msg]);
+    submitChat(input);
     setInput('');
   };
 
@@ -199,7 +113,10 @@ const Chat = () => {
               textColor={COLOR.primary}
               padding="0.2rem 0.2rem"
               fontSize="1.2rem"
-              onClick={submitChat}
+              onClick={() => {
+                submitChat(input);
+                setInput('');
+              }}
               disabled={locationState !== 1 && !isPending}
             />
           }
